refactor: dto 이동으로 인한 수정
